Refactor general settings code
